:new: pd controller with variable coefficients(2)
kd has three phases
kp has two phases ;in each phase it is varying.
result of test for this controller is shown in 2.fig
this controller has the least overshoot and has a short enough settling
time (needs improvement)
in summery  it is the most appropriate controller among these
controllers. :ok_hand: